+++ client/components/productList.js
@@ -3,7 +3,7 @@ import {connect} from 'react-redux';
 import {Link} from 'react-router-dom';
 import {allProducts, addProductThunk } from '../store';
 import ProductItem from './productItem';
-
+import NewProductForm from './newProduct'
 
 //logic to add a product will go here 
 
@@ -16,6 +16,9 @@ import ProductItem from './productItem';
 
   render () {
     const theProducts = this.props.products;
+
+    const { currentUser } = this.props;
+    const isAdmin = !!currentUser.isAdmin
 
     if (!theProducts.length) return  <div>No products found</div> 
     return (
@@ -30,92 +33,10 @@ import ProductItem from './productItem';
 
 
         </div>
-          <form onSubmit={ this.props.handleClick } >
-        <div>
-            <label htmlFor="artist">
-              <small>Product Artist</small>
-            </label>
-            <input
-              onChange={this.handleChange}
-              value={this.props.artist}
-              name="artist"
-              type="text"
-            />
-            </div>
-            <div>
-            <label htmlFor="description">
-              <small>Description</small>
-            </label>
-            <input
-              onChange={this.handleChange}
-              value={this.props.description}
-              name="description"
-              type="text"
-            />
-          </div>
-          <div>
-            <label htmlFor="imgUrl">
-              <small>imageUrl</small>
-            </label>
-            <input
-              onChange={this.handleChange}
-              value={this.props.imgUrl}
-              name="imgUrl"
-              type="text"
-            />
-          </div>
-          <div>
-            <label htmlFor="price">
-              <small>Price</small>
-            </label>
-            <input
-              onChange={this.handleChange}
-              value={this.props.price}
-              name="price"
-              type="number"
-              step="1"
-            />
-          </div>
-          <div>
-            <label htmlFor="quantity">
-              <small>Product Quantity</small>
-            </label>
-            <input
-              onChange={this.handleChange}
-              value={this.props.quantity}
-              name="quantity"
-              type="text"
-            />
-          </div>
+        {isAdmin && <NewProductForm /> }
+        </div>
 
 
-          <div>
-            <label htmlFor="title">
-              <small>Product Title</small>
-            </label>
-            <input
-              onChange={this.handleChange}
-              value={this.props.title}
-              name="title"
-              type="text"
-            />
-          </div>
-          <div>
-            <label htmlFor="title">
-              <small>Product Category</small>
-            </label>
-            <input
-              onChange={this.handleChange}
-              value={this.props.categoryId}
-              name="categoryId"
-              type="number"
-            />
-          </div>
-        <button type="submit">
-          Add a Product
-        </button>
-          </form>
-      </div>
     );
   }
 }
@@ -123,26 +44,20 @@ import ProductItem from './productItem';
 /**
  * CONTAINER
  */
-const mapState = ({products}) => ({products})
+// const mapState = ({products}) => ({products})
+
+const mapState = state => ({
+  currentUser: state.user,
+  products: state.products
+
+});
+
+
+
 
 const mapDispatch = dispatch => ({
   getAllProducts: () => {
       dispatch(allProducts())
-  },
-  handleClick: (event) => {
-    event.preventDefault()
-    const newProduct = {
-      artist: event.target.artist.value,
-      title: event.target.title.value,
-      description: event.target.description.value,
-      price: +event.target.price.value,
-      quantity: +event.target.quantity.value,
-      imgUrl: event.target.imgUrl.value,
-      categoryId: +event.target.categoryId.value
-    }
-    console.log("newProduct", newProduct)
-    console.log("addProductThunk", addProductThunk)
-    dispatch(addProductThunk(newProduct))
   }
 })
